# apps/docs/static/openapi/framework-live.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/me": {
      "get": {
        "operationId": "UserController_getCurrentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns current user profile."
          }
        },
        "summary": "Get current user",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UserController_updateCurrentUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostUserBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Current user updated."
          }
        },
        "summary": "Update current user",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UserController_deleteCurrentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user deleted."
          }
        },
        "summary": "Delete current user",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UserController_getUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns user details."
          }
        },
        "summary": "Get user by id",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UserController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostUserBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated."
          }
        },
        "summary": "Update user by id",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UserController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted."
          }
        },
        "summary": "Delete user by id",
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/users/me/customers/{id}": {
      "get": {
        "operationId": "UserController_getCustomerForCurrentUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Customer identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns customer details for current user."
          }
        },
        "summary": "Get current user customer by id",
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/products": {
      "get": {
        "operationId": "ProductsController_getProductList",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Product type filter.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Product category filter.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "description": "Optional locale code used to localize response values.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Sort expression from query string, for example `name_ASC`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "basePath",
            "required": false,
            "in": "query",
            "description": "Optional base path override used by integration routing.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns product list."
          }
        },
        "summary": "List products",
        "tags": [
          "products"
        ]
      }
    },
    "/api/products/{id}": {
      "get": {
        "operationId": "ProductsController_getProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": false,
            "in": "path",
            "description": "Optional variant identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": false,
            "in": "path",
            "description": "Optional locale code used to localize response values.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "basePath",
            "required": false,
            "in": "path",
            "description": "Optional base path override used by integration routing.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantOptionGroups",
            "required": false,
            "in": "path",
            "description": "Configuration of variant option groups used by product details views.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns product details."
          }
        },
        "summary": "Get product by id",
        "tags": [
          "products"
        ]
      }
    },
    "/api/products/{id}/variants/{variantId}/related-products": {
      "get": {
        "operationId": "ProductsController_getRelatedProductList",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "description": "Relation type to resolve (spare part, replacement, compatible service).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Source product identifier (bound from route `:id`).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "description": "Optional source product variant identifier (bound from route `:variantId`).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": false,
            "in": "path",
            "description": "Optional locale code used to localize response values.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "path",
            "description": "Sort expression from query string, for example `createdAt_DESC`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "basePath",
            "required": false,
            "in": "path",
            "description": "Optional base path override used by integration routing.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns related product list."
          }
        },
        "summary": "List related products",
        "tags": [
          "products"
        ]
      }
    },
...
    "/api/payments/providers": {
      "get": {
        "operationId": "PaymentsController_getProviders",
        "parameters": [
          {
            "name": "regionId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns available payment providers."
          }
        },
        "summary": "List payment providers",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/payments/sessions": {
      "post": {
        "operationId": "PaymentsController_createSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment session created."
          }
        },
        "summary": "Create payment session",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/payments/sessions/{id}": {
      "get": {
        "operationId": "PaymentsController_getSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns payment session details."
          }
        },
        "summary": "Get payment session by id",
        "tags": [
          "payments"
        ]
      },
      "patch": {
        "operationId": "PaymentsController_updateSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSessionBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment session updated."
          }
        },
        "summary": "Update payment session",
        "tags": [
          "payments"
        ]
      },
      "delete": {
        "operationId": "PaymentsController_cancelSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment session canceled."
          }
        },
        "summary": "Cancel payment session",
        "tags": [
          "payments"
        ]
      }
    },
...
  },
  "info": {
    "title": "O2S API Harmonization HTTP API",
    "description": "Live API documentation generated from current NestJS controllers and DTOs.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PostUserBody": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "User first name."
          },
          "lastName": {
            "type": "string",
            "description": "User last name."
          }
        },
        "required": [
          "firstName",
          "lastName"
        ]
      },
...
      "CreateSessionBody": {
        "type": "object",
        "properties": {
          "cartId": {
            "type": "string"
          },
          "providerId": {
            "type": "string"
          },
          "returnUrl": {
            "type": "string"
          },
          "cancelUrl": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          }
        },
        "required": [
          "cartId",
          "providerId",
          "returnUrl"
        ]
      },
      "UpdateSessionBody": {
        "type": "object",
        "properties": {
          "returnUrl": {
            "type": "string"
          },
          "cancelUrl": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          }
        }
      },
...
    }
  }
}